# api/openapi.yaml
openapi: 3.1.0
info:
  title: AgentOps Platform API
  version: 0.1.0
  description: |
    Control-plane API for operating AI agents with DevOps discipline:
    evaluation, regression testing, canary release and automatic rollback.

    The contract is framework-agnostic HTTP. The reference implementation
    targets agents built with Google ADK and deployed on Cloud Run.

    Evaluation axes:
    - `drift`      — output drift score (Gemini-as-judge score delta vs baseline)
    - `trajectory` — behavior evaluation (ADK Eval trajectory / tool-call sequence)
    - `cost`       — cost per task (derived from Cloud Trace / Logging)
    - `latency`    — p50/p95 latency (derived from Cloud Trace)
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://{service}.run.app/v1
    variables:
      service:
        default: agentops-platform
security:
  - bearerAuth: []
...
paths:
...
  /agents/{agentId}/deployments:
    parameters:
      - $ref: '#/components/parameters/agentId'
    post:
      tags: [deployments]
      operationId: createDeployment
      summary: Deploy a version with a canary strategy and rollback policy
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeploymentCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deployment' }
    get:
      tags: [deployments]
      operationId: listDeployments
      summary: List deployments
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Deployment' }
  /deployments/{deploymentId}:
    parameters:
      - $ref: '#/components/parameters/deploymentId'
    get:
      tags: [deployments]
      operationId: getDeployment
      summary: Get deployment status (canary ratio, health, state)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deployment' }
        '404': { $ref: '#/components/responses/NotFound' }
  /deployments/{deploymentId}:promote:
    parameters:
      - $ref: '#/components/parameters/deploymentId'
    post:
      tags: [deployments]
      operationId: promoteDeployment
      summary: Promote the canary to 100% traffic
      responses:
        '200':
          description: Promoted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deployment' }
  /deployments/{deploymentId}:rollback:
    parameters:
      - $ref: '#/components/parameters/deploymentId'
    post:
      tags: [deployments]
      operationId: rollbackDeployment
      summary: Roll back to the previous stable version
      description: |
        Also invoked automatically by the platform when the rollback policy
        thresholds are breached during the canary window.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200':
          description: Rolled back
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deployment' }
  /deployments/{deploymentId}:decide:
    parameters:
      - $ref: '#/components/parameters/deploymentId'
    post:
      tags: [deployments]
      operationId: decideDeployment
      summary: Run one meta-agent decision cycle
      description: |
        Executes one complete MetaAgentCycle for the given canary deployment.
        Evaluates the safety floor (Layer 1) and gray-zone signals (Layer 2)
        to decide whether to advance, hold, or rollback the canary.
        Returns a DecisionRecord as an audit trail entry.
      responses:
        '200':
          description: Decision cycle completed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DecisionRecord' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Deployment is already in a terminal state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Google-signed ID token (Cloud Run service-to-service auth).
  parameters:
    agentId:
      name: agentId
      in: path
      required: true
      schema: { type: string }
    deploymentId:
      name: deploymentId
      in: path
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    RollbackPolicy:
      type: object
      description: Auto-rollback triggers evaluated during the canary window
      properties:
        windowMinutes:
          type: integer
          description: Observation window after each traffic step
        maxDriftDrop:
          type: number
          description: Max allowed drop of drift score vs stable version
        maxTrajectoryDrop: { type: number }
        maxCostIncreaseRatio: { type: number }
        maxLatencyP95Ms: { type: number }
    DeploymentCreate:
      type: object
      required: [versionId, strategy]
      properties:
        versionId: { type: string }
        strategy:
          type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [canary, all-at-once]
            steps:
              type: array
              description: Traffic percentages for successive canary steps
              items: { type: integer, minimum: 1, maximum: 100 }
        rollbackPolicy: { $ref: '#/components/schemas/RollbackPolicy' }
    Deployment:
      allOf:
        - $ref: '#/components/schemas/DeploymentCreate'
        - type: object
          required: [deploymentId, state]
          properties:
            deploymentId: { type: string }
            state:
              type: string
              enum: [pending, canary, promoted, rolled_back, failed]
            currentTrafficPercent: { type: integer }
            previousStableVersionId: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...
    GrayZoneSignal:
      type: object
      description: Soft-limit breach signals used for gray-zone decisions
      properties:
        drift_drop:
          type: number
          description: Observed drift drop (stable − canary)
          default: 0.0
        trajectory_drop:
          type: number
          description: Observed trajectory drop
          default: 0.0
        cost_increase_ratio:
          type: number
          description: Cost increase ratio (0..N)
          default: 0.0
        canary_latency_ms:
          type: number
          description: Canary p95 latency in ms
          default: 0.0
        outcome_metrics:
          type: object
          additionalProperties: { type: number }
          description: Ingested outcome metric averages keyed by name
    DecisionRecord:
      type: object
      required: [decisionId, deploymentId, action, rationale, signal, judgedBy, decidedAt]
      description: Audit record of a single meta-agent decision cycle
      properties:
        decisionId: { type: string }
        deploymentId: { type: string }
        evaluationId: { type: string }
        action:
          type: string
          enum: [advance, hold, rollback]
        rationale:
          type: string
          description: Human-readable explanation of the decision
        signal: { $ref: '#/components/schemas/GrayZoneSignal' }
        judgedBy:
          type: string
          enum: [safety_floor, gemini, gemini_failed, heuristic, auto_advance, missing_baseline]
          description: Provenance of the decision
        prDraftId:
          type: string
          description: ID of the associated PRDraft when action==rollback
        decidedAt: { type: string, format: date-time }
...